Add sdake to environment working group leads
During the October 12th, 2018 TOC meeting, working group leads
were reviewed.  Costin had asked me via slack if I was interested
in the role of wg lead.  I had a conflict at the scheduled TOC
meeting time and the notes are a little dry as to decisions.  I
can't tell if this nominaition was approved from the notes.

I've submitted this PR to add myself as a co-lead to determine if
the nomination was approved.

Cheers
-steve