Add a new Triager role.
This leverages the new permissions introduced by GitHub as defined
in https://help.github.com/en/articles/repository-permission-levels-for-an-organization.

- The Triager role is assigned Triage permission and is used by project management
folks that don't contribute code but need to perform triage.

- The memnber nad maintainer role permissions are reduced to Triage. This allows
these roles to manipulate issues and PRs, but not to push to the repo. This is OK since
pushing is handled by bots.