add jasonwzm ux maintainers (#269)
jasonwzm is a config maintainer who has been responsible for a lot of
Istio's API tooling and versioning support. This type of work falls
under the charter of the merged Config+UX group. This PR is
essentially moving him from being a config maintainer to ux
maintainer. The config maintainer group will be removed when it no
longer owns any code via CODEOWNERS.