Add sdake to the T&R maintainers.
sdake is a Doc and Environment maintainer as well as a lead for the
Environemnt WG. Steve undertook a large portion of the work related to
builds using a buld container which would fall under the T&R umbrella.

Comments in PR# 20605 shows support from the other T&R WG leads as well
as Lin.