Create language about special Community Seat elections
In the 2021/07/02 Steering Committee, we discussed what happens if the company of a holder of a Community Seat earns a Contribution Seat. We raised the idea of holding a special election in ~Feb, immediately following the appointment of the Contribution Seat. This PR is language that describes this.

As this is a change to the Steering charter, it will require an affirmative vote of at least 80% of the Seats, or 11 of the 13.