Introduce `ztunnel` team (#1047)
This is purely for `CODEOWNERS` creation. We want a smallish group as
approvers for PRs. Using the full networking group is not ideal since it
spams ~20 folks on each PR, when the actual group involved in the
project is much smaller. Additionally, the expertise for reviewing
ztunnel PRs is quite different from other istio networking tasks.

The list was just informed by what I see as the most active reviewers
currently and is inteded as a starting place.