Update members.yaml (#1069)
The most common reason to send a PR to this repository is to join the Istio org. If you are joining, please fill out this template. If you are not, please delete all the below.

 I have reviewed the community membership guidelines
 I have reviewed the contribution guidelines
 I have enabled 2FA on my GitHub account
 I have joined the Istio discussion board and subscribed to the Contributors category
 I have joined Istio's Slack workspace
List your company name, or indicate Individual if you're not affiliated with a company:
DaoCloud
Provide a link to at least one PR that you have successfully pushed to one of the Istio repos:
istio/istio.io#12945