Adjust timezone link for more consistent results
in the Community Meeting section of the README, a link is given
to map the meeting's time to your local timezone.  This link uses a
query parameter to a google search, of "1000 am in pst".  Google does
a timezone conversion, but the wording implies we want 100am local time
in PST, which is inaccurate.  Removing the "in" results in the search
result we expect.