Remove requirement that TOC candidates be a Working Group Lead. (#1350)
* Remove requirement that TOC candidates be a Working Group Lead.

As the status of Istio Working Groups is a bit up in the air, this removes the requirement that an eligible TOC candidate must be a Working Group Lead.  No new requirements are put in place, but some guidance is given that new members should have made broad, sustained technical contributions over time.

* Update TECH-OVERSIGHT-COMMITTEE.md

Added working group lead requirements.

* Update TECH-OVERSIGHT-COMMITTEE.md

Update language.